Updated Submarine Controller with control locks
Added following public methods:
SetThrottleLock
SetRudderLock
SetBuoyancyLock

All take a boolean, with true indicating a lock is desired, and false desiring a lock to be lifted.
Locking a setting keeps that setting in its last requested target value, with the exception of throttle, which locks to zero.

If you desire more precise throttle locks I will happily change this, but this was the requested functionality.